+++ routes/monitorRoutes.js
@@ -1,6 +1,6 @@
-const express = require('express');
-const router = express.Router();
-const db = require('../db'); // Assumindo que o arquivo db exporta a função `createConnection`
+var express = require('express');
+var router = express.Router();
+var db = require('../db'); // Assumindo que o arquivo db exporta a função `createConnection`
 
 // Rota para obter o perfil do monitor
 router.get('/perfil/:id', async (req, res) => {
@@ -28,53 +28,8 @@ router.get('/perfil/:id', async (req, res) => {
     }
 });
 
-// Rota para criar uma nova sala
-router.post('/salas', async (req, res) => {
-    const { nome, capacidade, localizacao } = req.body;
 
-    if (!nome || !capacidade || !localizacao) {
-        return res.status(400).json({ message: 'Nome, capacidade e localização são obrigatórios.' });
-    }
 
-    try {
-        const conn = await db.createConnection(); // Criação da conexão aqui
-        await conn.query(
-            `INSERT INTO salas_de_aula (nome, capacidade, localizacao) VALUES (?, ?, ?)`,
-            [nome, capacidade, localizacao]
-        );
-        
-        conn.end(); // Fechando a conexão
-
-        res.status(201).json({ message: 'Sala criada com sucesso!' });
-    } catch (error) {
-        console.error('Erro ao criar sala:', error.message);
-        res.status(500).json({ message: 'Erro ao criar sala' });
-    }
-});
-
-// Rota para criar uma nova disciplina
-router.post('/disciplinas', async (req, res) => {
-    const { nome, descricao, monitorId } = req.body;
-
-    if (!nome || !descricao || !monitorId) {
-        return res.status(400).json({ message: 'Nome, descrição e ID do monitor são obrigatórios.' });
-    }
-
-    try {
-        const conn = await db.createConnection(); // Criação da conexão aqui
-        await conn.query(
-            `INSERT INTO disciplinas (nome, descricao, monitor_id) VALUES (?, ?, ?)`,
-            [nome, descricao, monitorId]
-        );
-        
-        conn.end(); // Fechando a conexão
-
-        res.status(201).json({ message: 'Disciplina criada com sucesso!' });
-    } catch (error) {
-        console.error('Erro ao criar disciplina:', error.message);
-        res.status(500).json({ message: 'Erro ao criar disciplina' });
-    }
-});
 
 // Rota para listar todas as disciplinas
 router.get('/disciplinas', async (req, res) => {
